requirements.txt
altair==6.1.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
confluent-kafka==2.14.0
databricks-sql-connector==4.2.6
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.47
idna==3.13
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lz4==4.4.5
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.4.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.1
pandas==2.3.3
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
pyarrow==24.0.0
pybreaker==1.4.1
pydeck==0.9.2
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
thrift==0.22.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
